backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
djangorestframework-requests
django-oauth-toolkit
google-cloud-documentai==2.6.2
openai==0.28
PyYAML==6.0
